+++ ContentPatcher/release-notes.md
@@ -2,7 +2,8 @@
 
 # Release notes
 ## Upcoming release
-* Content pack creators can now [validate `manifest.json` and `content.json`](https://github.com/Pathoschild/StardewMods/tree/develop/ContentPatcher#validate-contentjson) for common issues.
+* Content pack creators can now [validate `manifest.json` and `content.json`](https://github.com/Pathoschild/StardewMods/tree/develop/ContentPatcher#validate-json-files) for common issues.
+* Added `patch export` console command, which lets you see what an asset looks like with all changes applied.
 
 ## 1.9.2
 Released 25 July 2019.